# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText Messaging API
  description: |
    WASAText is a modern messaging platform that enables users to communicate
    through text and photo messages in both private conversations and group chats.
    
    This API provides endpoints for user authentication, message management,
    group operations, and conversation handling.
  version: "1.0"
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User identifier obtained from the login endpoint. 
        Include it in the Authorization header as: Bearer <identifier>
  schemas:
...
    UserProfile:
      type: object
      description: User profile information
      required:
        - identifier
        - username
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 128
          description: Unique user identifier
          example: "usr-a1b2c3d4"
        username:
          type: string
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
          description: User's display name
          example: "Alice"
        photoUrl:
          type: string
          format: uri
          minLength: 0
          maxLength: 500
          description: URL to user's profile photo
          example: "/photos/user-photo.jpg"
...
    GroupInfo:
      type: object
      description: Information about a group
      required:
        - groupId
        - groupName
        - participants
      properties:
        groupId:
          type: string
          minLength: 1
          maxLength: 128
          description: Unique group identifier
        groupName:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the group
        photoUrl:
          type: string
          format: uri
          minLength: 0
          maxLength: 500
          description: Group photo URL
        participants:
          type: array
          description: List of group members
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/UserProfile'
    ErrorResponse:
      type: object
      description: Error response format
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 500
          description: Human-readable error message
          example: "User not found"
paths:
...
  /groups/{groupId}/photo:
    put:
      tags: [group]
      summary: Update group photo
      description: Upload a new photo for the group
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group identifier
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: Photo file (JPEG, PNG, or GIF)
      responses:
        '200':
          description: Group photo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a group member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'